# Configure edm builder for dispatch only if there is tunneling