--- a/tt_metal/impl/dispatch/topology.cpp
+++ b/tt_metal/impl/dispatch/topology.cpp
@@ -1421,7 +1421,8 @@ void build_tt_fabric_program(
             tt::tt_metal::MetalContext::instance().get_cluster().get_associated_mmio_device(device->id());
         auto tunnels_from_mmio =
             tt::tt_metal::MetalContext::instance().get_cluster().get_devices_controlled_by_mmio_device(mmio_device_id);
-        // results are inclusive of the mmio_device_id
+        // results are inclusive of the mmio_device_id so they will never be zero
+        TT_ASSERT(tunnels_from_mmio.size() > 0);
         return (tunnels_from_mmio.size() - 1) > 0;
     }();