### Ticket
[Link to Github Issue](https://github.com/tenstorrent/tt-metal/issues/25118)

### Problem description
- Fixes 6U perf regression

### What's changed
- We should only configure the EDM Builder for dispatch only if the device is participating in a dispatch tunnel

Before:
Iteration : 0, Prefill Iteration : 0, tok/s/user : 191.26, Throughput : 6120.28 tok/s, Iteration Time : 5.23 ms
80L T/s/u estimate: 64.58306858713429

After:
Iteration : 0, Prefill Iteration : 0, tok/s/user : 284.32, Throughput : 9098.27 tok/s, Iteration Time : 3.52 ms
80L T/s/u estimate: 66.42449839842752

### Checklist
APC
https://github.com/tenstorrent/tt-metal/actions/runs/16327087096
T3K
https://github.com/tenstorrent/tt-metal/actions/runs/16327090741
T3K Unit
https://github.com/tenstorrent/tt-metal/actions/runs/16328342742
TG
https://github.com/tenstorrent/tt-metal/actions/runs/16327092161
TG Unit
https://github.com/tenstorrent/tt-metal/actions/runs/16328341561